CLI: refactor, move helpers related stuff to its own module
